+++ src/commonMain/kotlin/com/algolia/search/transport/internal/Transport.kt
@@ -5,17 +5,20 @@ import com.algolia.search.configuration.Compression
 import com.algolia.search.configuration.Configuration
 import com.algolia.search.configuration.Credentials
 import com.algolia.search.configuration.RetryableHost
+import com.algolia.search.exception.UnreachableHostsException
 import com.algolia.search.transport.RequestOptions
+import io.ktor.client.features.HttpRequestTimeoutException
 import io.ktor.client.features.ResponseException
+import io.ktor.client.features.timeout
 import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.request
 import io.ktor.http.HttpMethod
 import io.ktor.http.URLProtocol
+import io.ktor.network.sockets.ConnectTimeoutException
+import io.ktor.network.sockets.SocketTimeoutException
 import io.ktor.utils.io.errors.IOException
-import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import kotlinx.coroutines.withTimeout
 import kotlin.math.floor
 
 internal class Transport(
@@ -76,28 +79,45 @@ internal class Transport(
         body: String? = null,
     ): T {
         val hosts = callableHosts(callType)
-        val timeout = requestOptions.getTimeout(callType)
+        val errors by lazy(LazyThreadSafetyMode.NONE) { mutableListOf<Throwable>() }
         val requestBuilder = httpRequestBuilder(httpMethod, path, requestOptions, body)
 
         for (host in hosts) {
             requestBuilder.url.host = host.url
             try {
-                return withTimeout((host.retryCount + 1) * timeout) {
-                    httpClient.request<T>(requestBuilder).apply {
-                        mutex.withLock { host.reset() }
-                    }
+                setTimeout(requestBuilder, requestOptions, callType, host)
+                return httpClient.request<T>(requestBuilder).apply {
+                    mutex.withLock { host.reset() }
                 }
-            } catch (exception: TimeoutCancellationException) {
-                mutex.withLock { host.hasTimedOut() }
-            } catch (exception: IOException) {
-                mutex.withLock { host.hasFailed() }
-            } catch (exception: ResponseException) {
-                val value = exception.response?.status?.value ?: 0
-                val isRetryable = floor(value / 100f) != 4f
-
-                if (isRetryable) mutex.withLock { host.hasFailed() } else throw exception
+            } catch (exception: Exception) {
+                errors += exception
+                when (exception) {
+                    is HttpRequestTimeoutException, is SocketTimeoutException, is ConnectTimeoutException -> mutex.withLock { host.hasTimedOut() }
+                    is IOException -> mutex.withLock { host.hasFailed() }
+                    is ResponseException -> {
+                        val value = exception.response.status.value
+                        val isRetryable = floor(value / 100f) != 4f
+                        if (isRetryable) mutex.withLock { host.hasFailed() } else throw exception
+                    }
+                    else -> throw exception
+                }
             }
         }
-        throw RuntimeException("Unreachable hosts.")
+
+        throw UnreachableHostsException(errors)
+    }
+
+    /**
+     * Set socket read/write timeout.
+     */
+    private fun setTimeout(
+        requestBuilder: HttpRequestBuilder,
+        requestOptions: RequestOptions?,
+        callType: CallType,
+        host: RetryableHost,
+    ) {
+        requestBuilder.timeout {
+            socketTimeoutMillis = requestOptions.getTimeout(callType) * (host.retryCount + 1)
+        }
     }
 }